Initializes moui view for iOS.
The moui view is implemented as a View class which inherits the NativeView class. The moui view is designed for creating custom view based on OpenGL. It setups the OpenGL context for drawing. To use it, simply subclass the View class
and implement the rendering logic in the Render() method.

The NativeView class is a bridge to the real view object at specific platform.
It holds a pointer to the real native view as the native_handle_ class member.

A MOOpenGLView class is implemented for iOS to setup the environment for rendering OpenGL graphics.

Signed-off-by: Olli Wang <olliwang@ollix.com>