Initializes moui view for Android.
The initialization for Android is much like the same one for iOS. However, some method implementations are not done yet for Android such as the NativeView::SetBounds() method.

The OpenGL view is implemented as Java class and can be instantiated by the View class through JNI. To setup the moui environment for Android. The activity on the Java side must call a native method to initializes the Application class.

Signed-off-by: Olli Wang <olliwang@ollix.com>