Initializes the environment for widgets.
Widget-related source files are put in the "moui/widgets" directory. A WidgetView class is implemented for holding and rendering widgets. It sets up the rendering environment and manages a nanovg context for graphic drawing.

All new widgets should inherit the Widget class as base class and implement the rendering logic in the Render() method. The Render() function got the nanovg context that limit the drawing space to its bounds.

Signed-off-by: Olli Wang <olliwang@ollix.com>