Implements the wrapper class for platform-specific window.
This commit implements the new Window class as a wrapper for platform-specific window. The main window of the app can be retrieved through the GetMainWindow() method. For iOS and Android, the default main window can be retrieved out of the box. However, for Mac, the BaseWindow::RegisterMainWindow() method must be called before the first time calling GetMainWindow().

Signed-off-by: Olli Wang <olliwang@ollix.com>