Support for determining which device category the app is currently running on.
This commit implements the new Platform class for providing various platform information.

Currently it only supports determining which device category the app is currently running on. For Mac, it's clearly desktop. For iPhone and iPad, it's phone and tablet respectively.

For android devices, the smallest screen width dp is used to determine whether the device is a tablet or a phone. The default value is 600 dp. However, it can be changed through the SetSmallestScreenWidthDpForTablet() method depending on the needs.

Signed-off-by: Olli Wang <olliwang@ollix.com>