Support for setting widget's bounds dynamically.
This commit adds new methods for setting and getting widget's x, y, width and height dynamically. Now it's possible to align the widget related to its parent and sets value in percentage.

The WidgetView::RenderWidget() method is also improved to render only visible widgets. Visible widget means it's not hidden or have value of 0 for either width or height.

Signed-off-by: Olli Wang <olliwang@ollix.com>