Implements NativeView::SetBounds() for Android.
This commit creates the NativeView.java file for implementing the NativeView class on the java side to make it easier for development and debugging.

The NativeView::SetBounds() is finally implemented for Android in this commit. It uses `android.widget.FrameLayout` to set the position and dimension for the specific view.

Signed-off-by: Olli Wang <olliwang@ollix.com>