Supports animation for View and Widget classes.
This commit implmenets the animation mechanism to update OpenGL views synchronized to the refresh rate of the display continuously. This update applies to all of Mac, iOS, and Android platforms.

StartAnimation() and StopAnimation() methods are added in both View and Widget classes to control the animation mechanism. In addition, calling StartAnimation() gurantees the rendering will occur at least once.

The mechanism of updating view sychronized to the refresh rate of display is implmeneted through CADisplayLink and CVDisplayLink for iOS and Mac, respectively. For Android, it's implmeneted through GLSurfaceView.setRenderMode().