Updates rules for platform-specific source files.
This commit updates the gyp file and file names for platform-specific source files.

The GYP file adopts the strategy described in GYP's documentation (http://goo.gl/Cpz2Gf) that adds all source files to the global list and excludes files that are not designed for a specific platform.

The file name of Objective-C source files is changed to adopt the C++ style as well for consistency. This is also a desired approach when mixing Objective-C and C++ code as described in http://goo.gl/xdnZYV