Fixes the bug of checking if a widget collides a point.
The Widget::CollidePoint() method was determining whether a widget collides a point based on its origin related to its parent. However, the passed `point` parameter is related to the origin of the corresponded widget view.  To fix this issue, the method now determines the result based on the widget's origin related to its root widget, and which origin should be the same as related to the corresponded widget view.