Fixes View::Redraw() for Mac and iOS.
The `View::Redraw()` method should guarantee the view will be updated in the next refresh cycle of the display. However, if `Redraw()` is called in widget's rendering-related methods, the corresponded widget view won't update one more time.

This commmit adds a `_needsRedraw` property to native OpenGL views to make sure the display link won't be deattached when received a redraw request.