Bugfixes for OS X Yosemite.
This commit fixes various bugs and simplifies the code. Detailed changes are described as follow.

`NSOpenGLPFAWindow` is removed from OpenGL's pxiel format attributes since it's deprecated in OS X v10.9.

The last attribute of the OpenGL's pixel format is casted from 0 instead of `nil`, or build error will occur.

Fixes the bug that none is returned when the first time calling the `openGLContext` method.

Removes ` _pixelFormat` from instance member and allocates it as a local variable in `openGLContext` because where is the only place that requires it.

Removes the overriden `lockFocus` method because it doesn't matter in Yosemite.