Code refactoring.
This commit does code refactoring at various aspects.

The image generated by `nvgCreateImageSnapshot()` does not flip vertically anymore because this operation is expensive. Instead, it adds the `NVG_IMAGE_FLIPY` flag to `nvgCreateImageRGBA()` to achieve the same effect.

Setter methods of `Label`, `PageControl`, `ProgressView` and `Widget` classes are changed to request for redraw whenever necessary, and setters for setting color is changed to use the `NVGcolor` type for parameters directly.

The `Label` class also adjusts the vertical offset to draw the text box in some situations. This is a workaround to make the rendered position of the text more as expected.

In addition, the `ProgressView` class does not render the track or bar if the corresponded color is set to transparent.