Implements the `Button` widget.
This commits implements a new `Button` class inherited from the `Control` class. The ability to bind render functions and render for different states are moved from the `Control` class to the `Button` class as well. This makes the `Control` class as a base class for all control widgets.

A new style for rendering the default highlighted state is added. Now there are two flavors to choose and can be set through `set_default_highlighted_style()`.

In addition to render states for a button, the button also has a managed title label for displaying text. The text and color of the label can be set for each state separately.

In order to make it easier to customize the button. The `Button` class also has `rendering_offset_` and `title_edge_insets_` properties for changing the position of rendering and title label.