Fixes the bug that scroller may not disappeared properly.
Originally the animation is controlled between `WidgetWillRender()` and `WidgetDidRender()` method. However, because caching is enabled for the `Scroller` class. The actual rendering is done before calling `WidgetWillRender()`, which causes the problem to render the scroller with wrong animation progress.

This commit changes to control animation between `WidgetViewWillRender()` and `WidgetViewDidRender()` methods. That guarantees the rendering is always done for the current animation progress. The `animation_progress_` property is also reset after animation finished.