Support for propagating event to responders.
This commit changes the mechanism of handling events. Originally there is only one responder that can handle received events. Now the received event can be propagated to the next responder as long as its `Widget::HandleEvent()` method returns `true`.

A new `WidgetView::UpdateEventResponders()` method is created for populating a list of responder candidates. `WidgetView::HandleEvent()` then iterates the candidates to handle received events.

Other widget classes such as `Control` and `ScrollView` are also updated to adapt the changes. In addition, the `Control` class starts ignoring events if its position is changed. This may happen when a `Control` widget is put inside a `ScrollView` widget.