Resolves the situation where FBO failed to create.
This commit updates the `Widget::BeginRenderbufferUpdates()` method to return the result of creating the framebuffer object. The `nanovg` submodule is also updated to fix the bug where framebuffer object and renderbuffer object are not restored properly if failed.

In addition, `Button::Render()`, `Button::UpdateFramebuffer()`, and `Widget::RenderDefaultFramebuffer()` methods are also updated to adapt this change.

The `nvgluBindFramebuffer()` prototype function is removed from `nanovg_hook.h` as well because it already fixed in the `nanovg` submodule .