Implmenets the LinearLayout widget.
This commit implements the `LinearLayout` widget that is used to arrange child widgets in either horizontal or vertical direction. The `Layout` class is created as the base class for implementing every particular layout. Besides, it's inherited from the `ScrollView` class, so if its child widgets occupy larger space than the layout itself, the layout becomes scrollable.

In addition to the layout classes. The `ScrollView` class is also changed to better integrate with layouts. Originally, the parent widget of widgets added to a scroll view or layout is the `ScrollView`'s `content_view_`. Now their parent is set to the scroll view or layout widget directly. `ScrollView::GetContentViewSize()` is also implemented for the `LinearLayout` class.