Updates to use `float` type whenever it should be.
This commit fixes a lot of variables to use the `float` type. Basically all of these variables are used to draw something in nanovg, and which accepts `float` type in the first class.

This commit also updates the `Scroller` class to not cache rendering, or it would slow down the performance.