Support for interacting all scroll views in the responder chain.
This commit adds support for interacting all scroll views in the responder chain. To implement this feature, the scroll view needs to lock the scroll direction and pass the scroll action to underlying scroll view whenever suitable.

The `Widget` class adds a new `responder_chain_identifier_` property for the `WidgetView` class to find the same type of widgets in the responder chain in the `WidgetView::HandleEvent()` method. The `ScrollView` class takes advantage of this property to pass the scroll action to underlying scroll views.

Two important properties `acceptable_scroll_direction_` and `locked_scroll_direction_` are added to the `ScrollView` class so it knows when to lock the direction and when to ignore the scroll action.

In addition, both `Widget::StartAnimation()` and `Widget::StopAnimation()` methods are fixed to avoid unpredictable behaviors when scrolling.