Fixes how ScrollView animates in physics.
The acceleration used to calculate the acceleration-related quantities in physics were wrong. The number should be positive when animating to left or up, and be negative on the opposite direction. This commit fixes the `CalculateDisplacement()` and `CalculateInitialVelocity()` functions to change the sign of the acceleration number accordingly.

In addition, the `StopScrollingGradually()` function is changed to respond the current scroll direction only.