Updates layout widgets to arrange children thorugh `managed_widgets_`.
This commit changes how layout widgets arrange their child widgets. Previously, a layout widget used `children()` to iterate its child widgets in `ArrangeChildren()`. Now it's changed to iterate `managed_widgets_` to get its child widgets and corresponded dimensions. In addition, the `scale` property is taking into account when populating the list of managed widgets, which means the `width` and `height` properties of the `ManagedWidget` type are pre-calculated based on the widget's scale value.