Removes the `context_` property from the `Widget` class.
This commit removes the `context_` property from the `Widget` class mainly because the context object isn't required anymroe when deleting a framebuffer through nanovg's `nvgluDeleteFramebuffer()`.

To adapt this change, the `BindRenderFunction()` methods repleace the use of `std::ref` with `std::placeholders` so the `context_` property is no more needed as well.

On the other hand, the `Widget::UpdateContext()` method is removed as there is no `context_` property to update anymore, and now the `Widget::ContextWillChange()` method is responsible for releasing corresponded nanovg objects.

Last, because all widgets belonged to the same `WidgetView` use the same nanovg context for rendering. The `Widget::set_widget_view()` method now decides when to call the `ContextWillChange()` method.