Fixes the mechanism of removing a widget from its parent.
This commit changes the `Widget::RemoveFromParent()` method to remove a widget from its real parent instead of its logical parent.

The reason to do this change is obviously. For example, the logical parent of widgets added to a scroll view is the scroll view itself, but the real parent is the scroll view's content view. Removing the widgets directly from the scroll view would cause problems because the scroll view doesn't actually have those child widgets.