Support for Mac with Retina display.
This commit fixes the issue that the widget view cannot be rendered properly on Mac with Retina display. The fix is generally following the sample code from Apple's "GLEssentials" project.

In short, the `MOOpenGLView` now inherites from `NSOpenGLView` instead of `NSView`, and the OpenGL context and the display link is initialized in the `prepareOpenGL` method instead of the constructor.