Support for getting snapshot of widgets.
This commit implemenets the new `Widget::GetSnapshot()` method to return the widget's snapshot image. To achieve this work, a `WidgetView::Render()` overloading method is created to accept any managed widget as the root widget for rendering. The `Widget` class takes advantage of this change to render the widget itself to a newly created framebuffer. Once done, the snapshot pixels are read out from the framebuffer, and the framebuffer is deleted immediately afterward.