Updates layout widgets to respect children's alignment and offset.
This commit fixes the issue that widget's alignment and offset are not respected once added to a layout widget. To remedy this situation, the `Layout` class is changed to wrap managed widgets in so-called "cells".

Each cell is simply a widget that reserves the required space for each managed widget, and the `ArrangeCells()` method now takes control of changing the bounds of each cell. In order to know what size for a managed widget should be reserved, a new `GetOccupiedSpace()` method is also added to the `Widget` class to do the calculation.