Support for updating `Label`'s width to fit text.
This commit implements the new `UpdateWidthToFitText()` method for the `Label` class to update the label's width to fit its text. This method uses nanovg's `nvgTextBounds()` function to determine the required minimum width for the text. However, neither the returned value nor the difference between `xmax` and `xmin` values of the `bounds` parameter can always give the accurate width. As a workaround, the value whichever is greater is used as the text's width.