Improves the implementation of the `ScrollView` class.
This commit changes the implementation of the `ScrollView` class heavily to simplify the code and provide better user experience. It also changes the `WidgetView::HandleEvent()` method to improve how event responders work with underlying `ScrollView` widgets. The following describes various changes made to the `ScrollView` class.

The deceleration to stop content view gradually now is specified as the new `deceleration_rate_` property. The actual deceleration is calculated by multiplying the rate by the desired initial velocity. Besides, the `kScrollVelocityFactor` has been removed to adapt this change.

As long as `enables_paging_` is `true`, the scroll view now always stops on multiples of the scroll view's bounds when the user scrolls. Also, the `always_scroll_to_next_page_` property now is set to `false` by default to adapt this change.

Improves the feels of animating the content view, including switching between pages, stopping the view gradually, and the bounce effects.

The `GetContentViewOffset()` and `SetContentViewOffset()` methods are preserved for outside use. The class now only deals with content view's "origin" internally.

The scroll velocity now is calculated based on the accepted directions. If either horizontal or vertical direction is accepted but not both, the displacement now is determined by the distance of the two event locations.

A new `always_scroll_both_directions_` property is added to allow to lock the scroll direction to a specific direction or not.

The `event_history_` won't be cleared every time the "up" or "cancel" events are received so `GetScrollVelocity()` can still work in a overriding method.

The `IsScrolling()` method is replaced by the `is_scrolling_` property and the value is managed by the `HandleEvent()` method.

The `Scroller` class is also changed to redraw itself only if it has to.

Some bugs are fixed for the `ResolveContentViewOrigin()` method.