Adds transition effect to buttons.
This commit implements the transition effect for the `Button` widget. The transition happens when a finger dragged into the bounds of the button or from within a button to outside it bounds. More specifically, the transition happens when the button's control state changed between "highlighted" and "normal" states, or between "highlighted state" and "selected" states.

To support this feature, `TransitionBetweenControlStates()` and `StopTransitioningBetweenControlStates()` methods are created to start or end the transition. And the new `RenderFramebufferForTransition()` method is responsible for rendering the button's state while transitioning.