Fixes how an event is handled between control and scroll view widgets.
This commit fixes the logic of how an event is handled between control and scroll view widgets in the `WidgetView::HandleEvent()` method.

Besides, the returned value of the `Control::HandleEvent()` method is also fixed to follow the rule defined in `Widget::HandleEvent()` that the returned value should indicate whether the received event would be propagated to the next responder.

To fix the returned value of the `Control::HandleEvent()` method, the `Control::HandleControlEvents()` method is changed to return a boolean value indicating whether any of the passed control events has been handled by registered action callbacks.