Improves the mechanism of redrawing widgets.
This commit changes the `Widget::Redraw()` method to a virtual method so a subclass can override this method to reset custom framebuffers or reload its contents.

In addition, a new `visible_widgets_` property is added to the `WidgetView` class to keep a list of widgets that are previously rendered in the `Render()` method.

A new `WidgetView::Redraw()` overload is also added to redraw the entire widget view only if the passed widget is in the `visible_widgets_` list.