Fixes the bug that button may look inconsistent in size while transitioning.
This commit fixes the bug that a button widget may look inconsistent in size while transitioning between different states. The bug is caused by different  width and height values passed to the `nvgBeginFrame()` function in `RenderFramebufferForControlState()` and `RenderFramebufferForTransition()` methods, and it's fixed by converting the width and height to integer values explicitly.