Implements table view widget.
This commit implements the table view widget in new `TableView` and `TableViewCell` classes.  Two additional `TableViewDataSources` and `TableViewDelegate` classes are created to serve as table view's delegate classes. The idea to build the table view follows how table view works in iOS SDK.

Currently, the table view supports multiple sections and the table view cell includes a text label and an image view. The separators displayed in the table view are drawn in a dedicated `layout_view_` widget.

This commit also adds four new methods to the `Widget` class for managing the view hierarchy. The `ScrollView` class overwrites these methods as well, so the change will take place in the content view.

Besides, a `Clock:: ExecuteCallbackOnMainThread()` overload is added to provide an option for executing a callback with a delay time. This overload helps the table view to highlight its cell with a delay time, so the highlighted effect won't appear when scrolling the table view quickly.