Implements activity indicator view widget.
This commit implements the activity indicator view widget in the `ActivityIndicatorView` class. The idea to build the widget follows the `UIActivityIndicatorView` class from iOS SDK but providing more customization options.

For example, it's possible to set the number of lines that consists the indicator, and the number of visible lines can be defined as well. The colors of the beginning line and the last line are configurable, so is the line index to apply the beginning color.