Improves the mechanism of handling events.
This commit improves the mechanism of handling events in various ways, in particular for the `MOOpenGLView` class designed for iOS. It prevents from calling the `View::ShouldHandleEvent()` method multiple times if the event is already accepted.

Also, the `WidgetView::HandleEvent()` method has a new logic to forward the received event to its responders, and the `responder_chain_identifier_` property of the `Widget` class is removed because the `HandleEvent()` method does not rely on it anymore.

Last, the `HandleEvent()` method of `Control`, `ScrollView`, and `TableView` classes are improved to do their jobs better.