Builds the environments for bridging platform-specific objects.
This commit creates a new `native` directory and moves all bridging classes into it. All New bridging classes added in the future should put in this directory as well. Besides, a new `NativeObject` class is created as the base class of all bridging classes. This base class extends the possibility of bridging not only native views but any native objects.