Fixes the issue that widget view may freeze.
This commit fixes the issue that widget view may freeze due to not managing the animation well. The `Widget` class updates to take consideration of the `paused_animation_` property when trying to pause or resume animation.

Besides, the `WidgetView::HandleEvent()` method updates to make sure effective responders are cancelled manually if not receiving `cancel` event.