Re-implmenets Android support.
This commit re-implements the Android support to fit latest Android Studio 3 with Gradle and CMake build system for JNI. All the GYP scripts are replaced by CMake scripts as well, so there is no need to maintain two versions for different platforms.

Besides, all the source files that are written in Java are rewritten in Kotlin for simplicity. And most JNI code is rewritten to fix the memory issue.

This commit also implements a lot of missing methods for the Android target. However, the implementation for Android is still incomplete in this commit.