Fixes lag on macOS.
This commit fixes lag by removing the `renderOnMainThread` method from `MOView` on macOS. This method was added for executing `render` on main thread since currently frame updates are scheduled by CVDisplayLink and which executes `render` in background thread. Need to find another way to run `render` on the main thread.